pyproject.toml
[project]
name = "astro-cosmos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.10.5",
    "astronomer-cosmos>=1.9.2",
    "dbt-core>=1.9.3",
    "dbt-duckdb>=1.9.2",
    "deltalake>=0.25.4",
    "duckdb>=1.2.1",
    "requests>=2.32.3",
    "ruff>=0.11.0",
]
